Refactor Stake state components
