+++ components/TutorialModal/index.js
@@ -1,11 +1,11 @@
-import { useLearnMoreModalToggle } from "states/Application";
-import { useTutorialModalOpen, useTutorialModalToggle } from "contexts/Stake";
 import { setCookie } from "nookies";
 import { Fragment } from "react";
 import { useStateList } from "react-use";
+import { useLearnMoreModalToggle } from "states/Application";
 import Chevron from "svgs/Chevron";
-import Button from "./button";
-import Modal, { CloseButton } from "./modal";
+import Button from "../button";
+import Modal, { CloseButton } from "../modal";
+import { useTutorialModalOpen, useTutorialModalToggle } from "./state";
 
 const TUTORIAL_VIEWS = ["FIRST", "SECOND", "THIRD", "FOURTH"];
 
@@ -137,3 +137,5 @@ const TutorialModal = () => {
 };
 
 export default TutorialModal;
+
+export { useTutorialModalOpen, useTutorialModalToggle };
