+++ components/WithdrawModal/index.js
@@ -1,38 +1,34 @@
+import { parseUnits } from "@ethersproject/units";
 import { useWeb3React } from "@web3-react/core";
-import { useDepositModalOpen, useDepositModalToggle } from "contexts/Stake";
 import useCurrentToken from "hooks/useCurrentToken";
+import useStakingContract from "hooks/useStakingContract";
 import useToast, { FLAVORS } from "hooks/useToast";
-import useTokenBalance from "hooks/useTokenBalance";
-import { stake } from "lib/contracts/stake";
 import { useEffect } from "react";
 import { useForm } from "react-hook-form";
 import handleTxError from "util/handleTxError";
-import { roundDown } from "util/numbers";
-import Button from "./button";
-import Input from "./input";
-import LabelPair from "./labelPair";
-import Modal, { ModalHeader } from "./modal";
+import Button from "../button";
+import Input from "../input";
+import LabelPair from "../labelPair";
+import Modal, { ModalHeader } from "../modal";
+import { useWithdrawModalOpen, useWithdrawModalToggle } from "./state";
 
-const DepositModal = ({ totalStake, rewardsMultiplier, onComplete }) => {
-  const { library, chainId } = useWeb3React();
+const WithdrawModal = ({ withdrawableStake, totalStake, onComplete }) => {
+  const { chainId, library } = useWeb3React();
 
-  const open = useDepositModalOpen();
-  const toggle = useDepositModalToggle();
+  const open = useWithdrawModalOpen();
+  const toggle = useWithdrawModalToggle();
 
   const {
     handleSubmit,
     register,
     watch,
     setValue,
-    errors,
     formState,
+    errors,
     reset,
   } = useForm();
 
-  useEffect(() => {
-    reset();
-    updateDaiBalance();
-  }, [open]);
+  useEffect(() => reset(), [open]);
 
   const { dirty, isSubmitting } = formState;
 
@@ -41,19 +37,17 @@ const DepositModal = ({ totalStake, rewardsMultiplier, onComplete }) => {
 
   const DAI = useCurrentToken("DAI");
 
-  const { data: daiBalance = 0.0, mutate: updateDaiBalance } = useTokenBalance(
-    DAI
-  );
-
-  const flooredDaiBalance = roundDown(daiBalance);
-
   const addToast = useToast();
+
+  const stakingContract = useStakingContract();
 
   const onSubmit = async (values) => {
     let hidePendingToast = () => {};
 
     try {
-      const tx = await stake(DAI, values.amount, library.getSigner(), chainId);
+      const amount = parseUnits(values.amount, 18).toString();
+
+      const tx = await stakingContract.unstake(DAI, amount);
 
       const { hide: hidePending } = addToast(
         FLAVORS.TX_PENDING(tx.hash, chainId)
@@ -87,13 +81,15 @@ const DepositModal = ({ totalStake, rewardsMultiplier, onComplete }) => {
     }
   };
 
+  const calculateNewTotalStake = Number(totalStake) - amount;
+
   const formatNewTotalStake = Number(
-    parseFloat(amount || 0) + parseFloat(totalStake)
+    calculateNewTotalStake > 0 ? calculateNewTotalStake : 0
   ).toFixed(2);
 
   return (
     <Modal isOpen={open} onDismiss={toggle}>
-      <ModalHeader title="Deposit" onDismiss={toggle} />
+      <ModalHeader title="Withdraw" onDismiss={toggle} />
       <form
         method="POST"
         onSubmit={handleSubmit(onSubmit)}
@@ -109,18 +105,19 @@ const DepositModal = ({ totalStake, rewardsMultiplier, onComplete }) => {
           chip="DAI"
           id="amount"
           step="0.01"
+          name="amount"
           type="number"
           label="Amount"
           className="mb-8"
           placeholder="0.00"
-          setMaxValue={flooredDaiBalance}
-          setMax={() => setValue("amount", flooredDaiBalance)}
+          setMaxValue={withdrawableStake}
+          setMax={() => setValue("amount", withdrawableStake)}
           error={errors.amount}
           ref={register({
             required: "Please fill out this field",
             max: {
-              value: flooredDaiBalance,
-              message: "Not enough DAI in your wallet",
+              value: withdrawableStake,
+              message: "Insuficient withdrawable stake",
             },
             min: {
               value: 0.01,
@@ -151,4 +148,6 @@ const DepositModal = ({ totalStake, rewardsMultiplier, onComplete }) => {
   );
 };
 
-export default DepositModal;
+export default WithdrawModal;
+
+export { useWithdrawModalOpen, useWithdrawModalToggle };
