+++ components/stakeTable.js
@@ -1,9 +1,8 @@
-import { useLearnMoreModalToggle } from "states/Application";
-import { useTrustModalToggle } from "contexts/Stake";
 import useIsMember from "hooks/useIsMember";
 import useTrustData from "hooks/useTrustData";
 import { Fragment, useMemo, useState } from "react";
 import { useSortBy, useTable } from "react-table";
+import { useLearnMoreModalToggle } from "states/Application";
 import Chevron from "svgs/Chevron";
 import Info from "svgs/Info";
 import { healthTip } from "text/tooltips";
@@ -13,6 +12,7 @@ import AddressModal, {
 } from "./AddressModal/addressModal";
 import Button from "./button";
 import HealthBar from "./healthBar";
+import { useTrustModalToggle } from "./TrustModal";
 
 const StakeTableEmptyState = () => {
   const toggleLearnMoreModal = useLearnMoreModalToggle();
